refactor out some utility classes
